Unnest the exceptions package
