+++ cs3client/auth.py
@@ -17,7 +17,7 @@ from cs3.gateway.v1beta1.gateway_api_pb2_grpc import GatewayAPIStub
 from cs3.rpc.v1beta1.code_pb2 import CODE_OK
 
 from .cs3client import CS3Client
-from .exceptions.exceptions import AuthenticationException, SecretNotSetException
+from .exceptions import AuthenticationException, SecretNotSetException
 from .config import Config
