+++ tests/test_app.py
@@ -12,7 +12,7 @@ import cs3.rpc.v1beta1.code_pb2 as cs3code
 from unittest.mock import Mock, patch
 import pytest
 
-from cs3client.exceptions.exceptions import (
+from cs3client.exceptions import (
     AuthenticationException,
     NotFoundException,
     UnknownException,
