+++ tests/test_file.py
@@ -13,7 +13,7 @@ from unittest.mock import Mock, patch
 import cs3.rpc.v1beta1.code_pb2 as cs3code
 
 from cs3client.cs3resource import Resource
-from cs3client.exceptions.exceptions import (
+from cs3client.exceptions import (
     AuthenticationException,
     NotFoundException,
     FileLockedException,
